1.完善 人脸识别 功能；

+++ lib/trustdevice_pro_plugin.dart
@@ -82,7 +82,7 @@ class TDRiskCaptchaCallback {
 
 
 class TDLivenessCallback {
-  final void Function(String seqId,int errorCode,String errorMsg,int score,String bestImageString,String livenessId) onSuccess;
+  final void Function(String seqId,int errorCode,String errorMsg,double score,String bestImageString,String livenessId) onSuccess;
   final void Function(String seqId,int errorCode,String errorMsg,String livenessId) onFailed;
 
   const TDLivenessCallback({
